requirements.txt
flask
requests
psycopg2-binary
flask-restplus